V2.6.5: roaming-app safety-net widened to 5min + torn-read fix + 3 demotions
Widened ROAM_RECONNECT_SAFETY_NET_US 30s->5min so the experimental WiFi
roaming app fully owns ordinary reconnects (a router reboot alone can take
1-2 min of its own retries); the old 30s window was observed colliding with
an in-progress reconnect on a live node, forcing a spurious extra
disconnect/reconnect. Investigating that surfaced a torn-read hazard the
V2.6.3 sweep missed (t_attempt_start_us, written cross-task, feeding a
sub-second diagnostic) — fixed with a dedicated spinlock rather than
demotion. A follow-up general sweep found three unrelated bounded-duration
int64_t locals (mqtt.c, http_server.c, ntp.c) narrowed to 32-bit.

+++ main/mqtt.c
@@ -600,12 +600,12 @@ void mqtt_publish_state(const main_status_t *st,
         log_ftp_get_stats(&f);
         APPEND(",\"ftp_ok\":%s",       f.have_last && f.last_ok ? "true" : "false");
         APPEND(",\"ftp_bytes\":%" PRIu32, f.last_bytes);
-        int64_t age_s = 0;
+        int32_t age_s = 0;
         if (f.have_last && f.last_at > 0) {
             time_t now = time(NULL);
-            if (now > (time_t)f.last_at) age_s = (int64_t)now - f.last_at;
+            if (now > (time_t)f.last_at) age_s = (int32_t)((int64_t)now - f.last_at);
         }
-        APPEND(",\"ftp_age_s\":%" PRId64, age_s);
+        APPEND(",\"ftp_age_s\":%" PRId32, age_s);
     }
 #endif
